add read_text_file() to utils, import in all files that use it (reduce code duplication)
