+++ app/tools/rubric_generator/tools.py
@@ -9,19 +9,10 @@ from langchain_core.output_parsers import JsonOutputParser
 from langchain_google_genai import GoogleGenerativeAI
 from langchain_google_genai import GoogleGenerativeAIEmbeddings
 from app.services.logger import setup_logger
+from app.tools.utils.tool_utilities import read_text_file
 
 logger = setup_logger(__name__)
 
-def read_text_file(file_path):
-    # Get the directory containing the script file
-    script_dir = os.path.dirname(os.path.abspath(__file__))
-
-    # Combine the script directory with the relative file path
-    absolute_file_path = os.path.join(script_dir, file_path)
-    
-    with open(absolute_file_path, 'r') as file:
-        return file.read()
-    
 class RubricGenerator:
     def __init__(self, args=None, vectorstore_class=Chroma, prompt=None, embedding_model=None, model=None, parser=None, verbose=False):
         default_config = {
@@ -183,4 +174,3 @@ class RubricOutput(BaseModel):
     grade_level: str = Field(..., description="The grade level for which the rubric is created")
     criterias: List[RubricCriteria] = Field(..., description="The grading criteria for the rubric")
     feedback: str = Field(..., description="the feedback provided by the AI model on the generated rubric")
-
